backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.27.0
python-multipart==0.0.22
pdfplumber==0.10.3
python-docx==1.1.0
